// brokerchain-1.0.0/src/library/zip/compress/type.ts
// auto generated by dev/system

import { Logger } from "../../../myutils/logger.js";

export interface Input {
    from_path: string;
    to_file: string;
    without_parent_dir?: boolean;
}

export interface OutputOk {}

export interface OutputFail {}

export interface Callback<R> {
    ok: (output: OutputOk) => R;
    fail: (err: Error) => R;
}

export function check_input<R>(plog: Logger, v: any, cb: { ok: () => R; fail: (err: Error) => R }): R {
    const log = plog.sub("check_input");
    log.variable("v", v);
    try {
        log.println("v must be object");
        if (typeof v !== "object" || v === null) {
            throw new Error("v is not object");
        }

        Object.keys(v).forEach((field) => {
            if (field === "from_path") return;
            if (field === "to_file") return;
            if (field === "without_parent_dir") return;
            throw new Error("v contains unknown field: " + field);
        });

        log.println("v.from_path must be string");
        if (typeof v.from_path !== "string") {
            throw new Error("v.from_path is not string");
        }

        log.println("v.to_file must be string");
        if (typeof v.to_file !== "string") {
            throw new Error("v.to_file is not string");
        }

        if (v.without_parent_dir !== undefined) {
            log.println("v.without_parent_dir must be boolean");
            if (typeof v.without_parent_dir !== "boolean") {
                throw new Error("v.without_parent_dir is not boolean");
            }
        }
    } catch (err) {
        log.error(err);
        return cb.fail(err);
    }
    // nothing wrong
    log.ok();
    return cb.ok();
}

// JSON stringify value before copy to handle some specific problem
// eg. Date Object probelm
export function copy_output_ok(v: any): OutputOk {
    if (v !== undefined) {
        const v_cloned = JSON.parse(JSON.stringify(v));
        return _copy_output_ok(v_cloned);
    } else {
        return _copy_output_ok(v);
    }
}

export function _copy_output_ok(v: any): OutputOk {
    if (typeof v === "object" && v !== null) {
        const obj = {};
        return obj;
    } else {
        return {};
    }
}

// backward compatible purpose
export type Output = OutputOk;
export const copy_output = copy_output_ok;